Migrate to `Profile` class
