+++ blinky/ble/src/main/java/no/nordicsemi/android/blinky/ble/BlinkyManager.kt
@@ -6,7 +6,6 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.TimeoutCancellationException
 import kotlinx.coroutines.withContext
 import no.nordicsemi.android.blinky.spec.Blinky
-import no.nordicsemi.android.blinky.spec.BlinkySpec
 import no.nordicsemi.android.blinky.spec.exception.BlinkyException
 import no.nordicsemi.kotlin.ble.client.android.CentralManager
 import no.nordicsemi.kotlin.ble.client.android.Peripheral
@@ -37,41 +36,40 @@ class BlinkyManager(
         //
         // Note, that in this implementation the profile is installed before creating the connection,
         // but these can be swapped.
-        peripheral.profile(
-            serviceUuid = BlinkySpec.SERVICE_UUID,
-            required = true,
-        ) { remoteService ->
-            val ledButtonService = LedButtonServiceImpl(remoteService, this)
-
-            // Give the user control over the Blinky.
-            try {
-                block(ledButtonService)
-            } catch (e: CancellationException) {
-                // Don't disconnect when services were invalidated.
-                if (e.cause !is InvalidAttributeException) {
-                    peripheral.disconnect()
-                }
-                throw e
-            }
-        }
+        val ledButtonProfile = LedButtonProfileImpl(block)
+        peripheral.profile(ledButtonProfile)
 
         // Initiate connection, if not connected already.
         try {
             centralManager.connect(peripheral)
         } catch (_: TimeoutCancellationException) {
             throw BlinkyException.Timeout()
+        } catch (e: CancellationException) {
+            // Rethrow.
+            throw e
         } catch (_: Exception) {
             throw BlinkyException.ConnectionFailed()
         }
 
         // Keep the coroutine alive until the peripheral disconnects.
         // This method returns the disconnection reason.
-        val reason = peripheral.awaitDisconnection()
-        if (reason == Reason.RequiredServiceNotFound) {
-            throw BlinkyException.NotSupported()
-        }
-        if (reason != Reason.Success) {
-            throw BlinkyException.LinkLoss()
+        try {
+            val reason = peripheral.awaitDisconnection()
+            if (reason == Reason.RequiredServiceNotFound) {
+                throw BlinkyException.NotSupported()
+            }
+            if (reason != Reason.Success) {
+                throw BlinkyException.LinkLoss()
+            }
+        } catch (e: CancellationException) {
+            // The scope may get canceled when user leaves the screen.
+            // In that case, make sure to disconnect.
+            // Don't disconnect when services were invalidated, as the profile will be re-launched.
+            if (e.cause !is InvalidAttributeException) {
+                peripheral.disconnect()
+            }
+            // Rethrow.
+            throw e
         }
     }
 }
